refactor: create function for common pagination logic
and rename pagination to offsetPagination
to prepare new cursor pagination
