+++ src/util/pagination.js
@@ -1,4 +1,4 @@
-function getPaginationMeta ({ nbResults, page, nbResultsPerPage }) {
+function getOffsetPaginationMeta ({ nbResults, page, nbResultsPerPage }) {
   let nbPages = Math.floor(nbResults / nbResultsPerPage)
   if (nbResults % nbResultsPerPage !== 0) {
     nbPages += 1
@@ -14,8 +14,46 @@ function getPaginationMeta ({ nbResults, page, nbResultsPerPage }) {
   return paginationMeta
 }
 
+async function offsetPaginate ({
+  queryBuilder,
+  orderBy,
+  order,
+  nbResultsPerPage,
+  page,
+  applyOrder = true,
+}) {
+  // Clone the query builder to have the count for all matched results before pagination filtering
+  const countQueryBuilder = queryBuilder.clone()
+
+  if (applyOrder) {
+    queryBuilder.orderBy(orderBy, order)
+  }
+
+  queryBuilder
+    .offset((page - 1) * nbResultsPerPage)
+    .limit(nbResultsPerPage)
+
+  const [
+    results,
+    [{ count: nbResults }]
+  ] = await Promise.all([
+    queryBuilder,
+    countQueryBuilder.count()
+  ])
+
+  const paginationMeta = getOffsetPaginationMeta({
+    nbResults,
+    nbResultsPerPage,
+    page
+  })
+
+  paginationMeta.results = results
+  return paginationMeta
+}
+
 module.exports = {
-  getPaginationMeta,
+  offsetPaginate,
+  getOffsetPaginationMeta,
 
   DEFAULT_NB_RESULTS_PER_PAGE: 20,
 }
